requirements.txt
2to3==1.0
alabaster==0.7.13
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
Babel==2.13.1
backcall==0.2.0
blessings==1.7
boto3==1.26.135
botocore==1.29.165
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
coverage==7.2.5
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.1
django-configurations==2.4.1
django-filter==23.2
django-model-utils==4.3.1
django-nose==1.4.7
django-storages==1.13.2
django-unique-upload==0.2.1
djangorestframework==3.14.0
docutils==0.20.1
executing==2.0.0
factory-boy==3.2.1
Faker==19.11.0
flake8==6.0.0
ghp-import==2.1.0
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
ipdb==0.13.13
ipython==8.13.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.3
mock==5.0.2
newrelic==8.8.0
nose==1.3.7
nose-progressive==1.5.2
nose-py3==1.6.3
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
watchdog==3.0.0
wcwidth==0.2.8
xlrd==2.0.1
xlwt==1.3.0
